+++ lib/shopify/src/shopify_customer.dart
@@ -1,6 +1,7 @@
 import 'package:shopify_flutter/graphql_operations/storefront/mutations/customer_address_create.dart';
 import 'package:shopify_flutter/graphql_operations/storefront/mutations/customer_address_delete.dart';
 import 'package:shopify_flutter/graphql_operations/storefront/mutations/customer_address_update.dart';
+import 'package:shopify_flutter/graphql_operations/storefront/mutations/customer_default_address_update.dart';
 import 'package:shopify_flutter/graphql_operations/storefront/mutations/customer_update.dart';
 import 'package:shopify_flutter/mixins/src/shopfiy_error.dart';
 import 'package:shopify_flutter/models/src/shopify_user/address/address.dart';
@@ -149,6 +150,25 @@ class ShopifyCustomer with ShopifyError {
     checkForError(
       result,
       key: 'customerAddressDelete',
+      errorKey: 'customerUserErrors',
+    );
+  }
+
+  /// updates the default adderess to the [addressId] provided from the customer to which [customerAccessToken] belongs to.
+  Future<void> customerDefaultAddressUpdate({
+    required String addressId,
+    required String customerAccessToken,
+  }) async {
+    final MutationOptions _options = MutationOptions(
+        document: gql(customerDefaultAddressUpdateMutation),
+        variables: {
+          'customerAccessToken': customerAccessToken,
+          'addressId': addressId
+        });
+    final QueryResult result = await _graphQLClient!.mutate(_options);
+    checkForError(
+      result,
+      key: 'customerDefaultAddressUpdate',
       errorKey: 'customerUserErrors',
     );
   }